router and main page in progress
