+++ frontend/src/index.js
@@ -2,29 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "mapbox-gl/dist/mapbox-gl.css";
 import "./index.css";
-import App from "./App";
+import Map from "./Map";
+import MainPage from "./MainPage";
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Map />
+    <MainPage />
   </React.StrictMode>,
   document.getElementById("root")
 );
-
-// import React from 'react';
-// import ReactDOM from 'react-dom/client';
-// import './index.css';
-// import App from './App';
-// import reportWebVitals from './reportWebVitals';
-//
-// const root = ReactDOM.createRoot(document.getElementById('root'));
-// root.render(
-//   <React.StrictMode>
-//     <App />
-//   </React.StrictMode>
-// );
-//
-// // If you want to start measuring performance in your app, pass a function
-// // to log results (for example: reportWebVitals(console.log))
-// // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
